code structure follow-up
